+++ golem-common/src/model/component_constraint.rs
