+++ golem-component-service-base/src/model.rs
@@ -1,5 +1,5 @@
 use golem_common::model::component_metadata::ComponentMetadata;
-use golem_common::model::constraint::{FunctionUsage, FunctionUsageCollection};
+use golem_common::model::component_constraint::{FunctionUsage, FunctionUsageCollection};
 use golem_common::model::{ComponentId, ComponentType};
 use golem_service_base::model::{ComponentName, VersionedComponentId};
 use rib::WorkerFunctionsInRib;
