+++ golem-component-service-base/src/service/component.rs
@@ -23,7 +23,7 @@ use crate::service::component_processor::process_component;
 use async_trait::async_trait;
 use chrono::Utc;
 use golem_common::model::component_metadata::ComponentProcessingError;
-use golem_common::model::constraint::FunctionUsageCollection;
+use golem_common::model::component_constraint::FunctionUsageCollection;
 use golem_common::model::{ComponentId, ComponentType};
 use golem_common::SafeDisplay;
 use golem_service_base::model::{ComponentName, VersionedComponentId};
