+++ golem-worker-service-base/src/service/component/default.rs
@@ -12,7 +12,7 @@ use golem_api_grpc::proto::golem::component::ComponentConstraints;
 use golem_api_grpc::proto::golem::component::FunctionUsageCollection as FunctionUsageCollectionProto;
 use golem_common::client::{GrpcClient, GrpcClientConfig};
 use golem_common::config::RetryConfig;
-use golem_common::model::constraint::FunctionUsageCollection;
+use golem_common::model::component_constraint::FunctionUsageCollection;
 use golem_common::model::ComponentId;
 use golem_common::retries::with_retries;
 use golem_service_base::model::Component;
