+++ golem-worker-service-base/tests/services_tests.rs
@@ -29,7 +29,7 @@ use golem_worker_service_base::service::http::http_api_definition_validator::{
 };
 
 use chrono::Utc;
-use golem_common::model::constraint::{FunctionUsage, FunctionUsageCollection};
+use golem_common::model::component_constraint::{FunctionUsage, FunctionUsageCollection};
 use golem_wasm_ast::analysis::analysed_type::str;
 use rib::WorkerFunctionsInRib;
 use std::sync::Arc;
