+++ golem-worker-service/src/api/api_definition.rs
@@ -308,7 +308,7 @@ mod test {
     use crate::service::component::ComponentService;
     use async_trait::async_trait;
     use golem_common::config::DbSqliteConfig;
-    use golem_common::model::constraint::FunctionUsageCollection;
+    use golem_common::model::component_constraint::FunctionUsageCollection;
     use golem_common::model::ComponentId;
     use golem_service_base::db;
     use golem_service_base::model::Component;
